analyze audio as it is collected

+++ SmartAlarm/src/com/example/smartalarm/AudioDataPoint.java
@@ -3,9 +3,9 @@ package com.example.smartalarm;
 public class AudioDataPoint {
 
 	public long time;
-	public short[] data;
+	public float[] data;
 	
-	public AudioDataPoint(long time, short[] data) {
+	public AudioDataPoint(long time, float[] data) {
 		this.time = time;
 		this.data = data;
 	}
